refactor: image 유틸 함수 리팩토링
