+++ src/utils/delay.ts
@@ -1,0 +1,3 @@
+export function delay(sec: number) {
+  return new Promise((resolve) => setTimeout(resolve, sec));
+}
